+++ utils/utils.h
@@ -16,7 +16,7 @@
 #endif
 
 uint32_t* generateKey();
-long getFileSize(FILE* in);
+size_t getFileSize(FILE* in);
 char* getEncFileName(const char* fileName);
 void GNN_teaEncrypt(uint32_t* GNN_block, const uint32_t* GNN_k);
 std::string getRestoredFileName(const char* oldFileName);
